@@ -20,7 +20,7 @@ namespace detail {
 template <typename Functor, typename... Args>
 std::enable_if_t<
     std::is_member_pointer<std::decay_t<Functor>>::value,
-    std::result_of_t<Functor&&(Args&&...)>>
+    std::invoke_result_t<Functor&&, Args&&...>>
 invoke(Functor&& f, Args&&... args)
 {
     return std::mem_fn(f)(std::forward<Args>(args)...);
@@ -29,7 +29,7 @@ invoke(Functor&& f, Args&&... args)
 template <typename Functor, typename... Args>
 std::enable_if_t<
     !std::is_member_pointer<std::decay_t<Functor>>::value,
-    std::result_of_t<Functor&&(Args&&...)>>
+    std::invoke_result_t<Functor&&, Args&&...>>
 invoke(Functor&& f, Args&&... args)
 {
     return std::forward<Functor>(f)(std::forward<Args>(args)...);
