@@ -712,9 +712,7 @@ bool ExecutionNode::doWalk(WalkerWorkerBase<ExecutionNode>& worker, bool subQuer
   // we can be quite generous with the buffer size since the implementation is not recursive.
   constexpr std::size_t NumBufferedEntries = 1000;
   constexpr std::size_t BufferSize = sizeof(Entry) * NumBufferedEntries;
-  using Container = containers::SmallVector<Entry, BufferSize>;
-  Container::allocator_type::arena_type arena;
-  Container nodes{arena};
+  containers::SmallVectorWithArena<Entry, BufferSize> nodes;
   // Our stackbased arena can hold NumBufferedEntries, so we reserve everythhing at once.
   nodes.reserve(NumBufferedEntries);
   nodes.emplace_back(const_cast<ExecutionNode*>(this), State::Pending);
@@ -2108,11 +2106,10 @@ bool SubqueryNode::mayAccessCollections() {
       ExecutionNode::SHORTEST_PATH,
       ExecutionNode::K_SHORTEST_PATHS};
 
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodes;
 
   NodeFinder<std::initializer_list<ExecutionNode::NodeType>, WalkerUniqueness::Unique> finder(
-      types, nodes, true);
+      types, nodes.vector(), true);
   _subquery->walk(finder);
 
   if (!nodes.empty()) {
