@@ -949,8 +949,8 @@ Collection* addCollectionToQuery(QueryContext& query, std::string const& cname,
 /// @brief adds a SORT operation for IN right-hand side operands
 void arangodb::aql::sortInValuesRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
                                      OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::FILTER, true);
 
   bool modified = false;
@@ -1139,8 +1139,8 @@ void arangodb::aql::sortInValuesRule(Optimizer* opt, std::unique_ptr<ExecutionPl
 void arangodb::aql::removeRedundantSortsRule(Optimizer* opt,
                                              std::unique_ptr<ExecutionPlan> plan,
                                              OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::SORT, true);
 
   if (nodes.empty()) {
@@ -1289,8 +1289,8 @@ void arangodb::aql::removeRedundantSortsRule(Optimizer* opt,
 void arangodb::aql::removeUnnecessaryFiltersRule(Optimizer* opt,
                                                  std::unique_ptr<ExecutionPlan> plan,
                                                  OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::FILTER, true);
 
   bool modified = false;
@@ -1341,11 +1341,8 @@ void arangodb::aql::removeUnnecessaryFiltersRule(Optimizer* opt,
 void arangodb::aql::removeCollectVariablesRule(Optimizer* opt,
                                                std::unique_ptr<ExecutionPlan> plan,
                                                OptimizerRule const& rule) {
-  using short_alloc_t = ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type;
-  auto arena = short_alloc_t::arena_type();
-  auto nodes = containers::SmallVector<ExecutionNode*>(arena);
-  // NOLINTNEXTLINE(bugprone-sizeof-expression)
-  nodes.reserve(short_alloc_t::size / sizeof(short_alloc_t::value_type));
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::COLLECT, true);
 
   bool modified = false;
@@ -1460,8 +1457,8 @@ class PropagateConstantAttributesHelper {
 
   /// @brief inspects a plan and propagates constant values in expressions
   void propagateConstants() {
-    ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-    ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+    ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+    auto& nodes = nodesStorage.vector();
     _plan->findNodesOfType(nodes, EN::FILTER, true);
 
     for (auto const& node : nodes) {
@@ -1693,15 +1690,15 @@ void arangodb::aql::propagateConstantAttributesRule(Optimizer* opt,
 void arangodb::aql::moveCalculationsUpRule(Optimizer* opt,
                                            std::unique_ptr<ExecutionPlan> plan,
                                            OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::CALCULATION, true);
 
   SmallUnorderedMap<ExecutionNode*, ExecutionNode*>::allocator_type::arena_type subqueriesArena;
   SmallUnorderedMap<ExecutionNode*, ExecutionNode*> subqueries{subqueriesArena};
   {
-    SmallVector<ExecutionNode*>::allocator_type::arena_type s;
-    SmallVector<ExecutionNode*> subs{s};
+    ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> subsStorage;
+    auto& subs = subsStorage.vector();
     plan->findNodesOfType(subs, ExecutionNode::SUBQUERY, true);
 
     // we build a map of the top-most nodes of each subquery to the outer subquery node
@@ -1811,8 +1808,8 @@ void arangodb::aql::moveCalculationsUpRule(Optimizer* opt,
 void arangodb::aql::moveCalculationsDownRule(Optimizer* opt,
                                              std::unique_ptr<ExecutionPlan> plan,
                                              OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, {EN::CALCULATION, EN::SUBQUERY}, true);
 
   std::vector<ExecutionNode*> stack;
@@ -1926,8 +1923,8 @@ void arangodb::aql::moveCalculationsDownRule(Optimizer* opt,
 void arangodb::aql::specializeCollectRule(Optimizer* opt,
                                           std::unique_ptr<ExecutionPlan> plan,
                                           OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::COLLECT, true);
 
   bool modified = false;
@@ -2070,8 +2067,8 @@ void arangodb::aql::specializeCollectRule(Optimizer* opt,
 /// filters are not pushed beyond limits
 void arangodb::aql::moveFiltersUpRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
                                       OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::FILTER, true);
 
   std::vector<ExecutionNode*> stack;
@@ -2167,8 +2164,8 @@ struct VariableReplacer final
 void arangodb::aql::simplifyConditionsRule(Optimizer* opt,
                                            std::unique_ptr<ExecutionPlan> plan,
                                            OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::CALCULATION, true);
 
   if (nodes.empty()) {
@@ -2396,8 +2393,8 @@ void arangodb::aql::simplifyConditionsRule(Optimizer* opt,
 /// @brief fuse filter conditions that follow each other
 void arangodb::aql::fuseFiltersRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
                                     OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::FILTER, true);
 
   if (nodes.size() < 2) {
@@ -2486,8 +2483,8 @@ void arangodb::aql::fuseFiltersRule(Optimizer* opt, std::unique_ptr<ExecutionPla
 void arangodb::aql::removeRedundantCalculationsRule(Optimizer* opt,
                                                     std::unique_ptr<ExecutionPlan> plan,
                                                     OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::CALCULATION, true);
 
   if (nodes.size() < 2) {
@@ -2613,8 +2610,8 @@ void arangodb::aql::removeRedundantCalculationsRule(Optimizer* opt,
 void arangodb::aql::removeUnnecessaryCalculationsRule(Optimizer* opt,
                                                       std::unique_ptr<ExecutionPlan> plan,
                                                       OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, ::removeUnnecessaryCalculationsNodeTypes, true);
 
   ::arangodb::containers::HashSet<ExecutionNode*> toUnlink;
@@ -2813,8 +2810,8 @@ void arangodb::aql::useIndexesRule(Optimizer* opt, std::unique_ptr<ExecutionPlan
                                    OptimizerRule const& rule) {
   // These are all the nodes where we start traversing (including all
   // subqueries)
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findEndNodes(nodes, true);
 
   std::unordered_map<ExecutionNodeId, ExecutionNode*> changes;
@@ -3234,8 +3231,8 @@ struct SortToIndexNode final
 
 void arangodb::aql::useIndexForSortRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
                                         OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::SORT, true);
 
   bool modified = false;
@@ -3258,8 +3255,8 @@ void arangodb::aql::useIndexForSortRule(Optimizer* opt, std::unique_ptr<Executio
 void arangodb::aql::removeFiltersCoveredByIndexRule(Optimizer* opt,
                                                     std::unique_ptr<ExecutionPlan> plan,
                                                     OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::FILTER, true);
 
   ::arangodb::containers::HashSet<ExecutionNode*> toUnlink;
@@ -3404,8 +3401,8 @@ static bool NextPermutationTuple(std::vector<size_t>& data, std::vector<size_t>&
 void arangodb::aql::interchangeAdjacentEnumerationsRule(Optimizer* opt,
                                                         std::unique_ptr<ExecutionPlan> plan,
                                                         OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
 
   plan->findNodesOfType(nodes, ::interchangeAdjacentEnumerationsNodeTypes, true);
 
@@ -3757,8 +3754,8 @@ void moveScatterAbove(ExecutionPlan& plan, ExecutionNode* at) {
 // Find all Subquery Nodes
 void arangodb::aql::findSubqueriesInPlan(ExecutionPlan& plan,
                                          SmallUnorderedMap<ExecutionNode*, ExecutionNode*>& subqueries) {
-  SmallVector<ExecutionNode*>::allocator_type::arena_type s;
-  SmallVector<ExecutionNode*> subs{s};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> subsStorage;
+  auto& subs = subsStorage.vector();
   plan.findNodesOfType(subs, ExecutionNode::SUBQUERY, true);
 
   for (auto& it : subs) {
@@ -3783,8 +3780,8 @@ void arangodb::aql::scatterInClusterRule(Optimizer* opt, std::unique_ptr<Executi
 
   // we are a coordinator. now look in the plan for nodes of type
   // EnumerateCollectionNode, IndexNode, IResearchViewNode, and modification nodes
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, ::scatterInClusterNodeTypes, true);
 
   TRI_ASSERT(plan->getAst());
@@ -4077,8 +4074,8 @@ void arangodb::aql::distributeInClusterRule(Optimizer* opt,
   // we are a coordinator, we replace the root if it is a modification node
 
   // only replace if it is the last node in the plan
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> subqueryNodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> subqueryNodesStorage;
+  auto& subqueryNodes = subqueryNodesStorage.vector();
   // inspect each return node and work upwards to SingletonNode
   subqueryNodes.push_back(plan->root());
   plan->findNodesOfType(subqueryNodes, ExecutionNode::SUBQUERY, true);
@@ -4178,8 +4175,8 @@ void arangodb::aql::collectInClusterRule(Optimizer* opt, std::unique_ptr<Executi
   TRI_ASSERT(arangodb::ServerState::instance()->isCoordinator());
   bool wasModified = false;
 
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::COLLECT, true);
 
   VarSet allUsed;
@@ -4484,8 +4481,8 @@ void arangodb::aql::distributeFilterCalcToClusterRule(Optimizer* opt,
                                                       OptimizerRule const& rule) {
   bool modified = false;
 
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::GATHER, true);
 
   VarSet varsSetHere;
@@ -4637,8 +4634,8 @@ void arangodb::aql::distributeFilterCalcToClusterRule(Optimizer* opt,
 void arangodb::aql::distributeSortToClusterRule(Optimizer* opt,
                                                 std::unique_ptr<ExecutionPlan> plan,
                                                 OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   VarSet usedBySort;
   plan->findNodesOfType(nodes, EN::GATHER, true);
 
@@ -4770,8 +4767,8 @@ void arangodb::aql::distributeSortToClusterRule(Optimizer* opt,
 void arangodb::aql::removeUnnecessaryRemoteScatterRule(Optimizer* opt,
                                                        std::unique_ptr<ExecutionPlan> plan,
                                                        OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::REMOTE, true);
 
   ::arangodb::containers::HashSet<ExecutionNode*> toUnlink;
@@ -4856,8 +4853,8 @@ void arangodb::aql::restrictToSingleShardRule(Optimizer* opt,
     return;
   }
 
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::REMOTE, true);
 
   ::arangodb::containers::HashSet<ExecutionNode*> toUnlink;
@@ -5256,8 +5253,8 @@ class RemoveToEnumCollFinder final
 void arangodb::aql::undistributeRemoveAfterEnumCollRule(Optimizer* opt,
                                                         std::unique_ptr<ExecutionPlan> plan,
                                                         OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, ::undistributeNodeTypes, true);
 
   ::arangodb::containers::HashSet<ExecutionNode*> toUnlink;
@@ -5558,8 +5555,8 @@ struct OrSimplifier {
 //  the same (single) attribute.
 void arangodb::aql::replaceOrWithInRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
                                         OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::FILTER, true);
 
   bool modified = false;
@@ -5727,8 +5724,8 @@ struct RemoveRedundantOr {
 void arangodb::aql::removeRedundantOrRule(Optimizer* opt,
                                           std::unique_ptr<ExecutionPlan> plan,
                                           OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::FILTER, true);
 
   bool modified = false;
@@ -5775,8 +5772,8 @@ void arangodb::aql::removeDataModificationOutVariablesRule(Optimizer* opt,
                                                            OptimizerRule const& rule) {
   bool modified = false;
 
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, ::removeDataModificationOutVariablesNodeTypes, true);
 
   for (auto const& n : nodes) {
@@ -5848,8 +5845,8 @@ void arangodb::aql::removeDataModificationOutVariablesRule(Optimizer* opt,
 void arangodb::aql::optimizeTraversalsRule(Optimizer* opt,
                                            std::unique_ptr<ExecutionPlan> plan,
                                            OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> tNodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> tNodesStorage;
+  auto& tNodes = tNodesStorage.vector();
   plan->findNodesOfType(tNodes, EN::TRAVERSAL, true);
 
   if (tNodes.empty()) {
@@ -5929,8 +5926,8 @@ void arangodb::aql::optimizeTraversalsRule(Optimizer* opt,
 
   if (!tNodes.empty()) {
     // These are all the end nodes where we start
-    ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-    ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+    ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+    auto& nodes = nodesStorage.vector();
     plan->findEndNodes(nodes, true);
 
     for (auto const& n : nodes) {
@@ -5946,8 +5943,8 @@ void arangodb::aql::optimizeTraversalsRule(Optimizer* opt,
 void arangodb::aql::removeFiltersCoveredByTraversal(Optimizer* opt,
                                                     std::unique_ptr<ExecutionPlan> plan,
                                                     OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> fNodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> fNodesStorage;
+  auto& fNodes = fNodesStorage.vector();
   plan->findNodesOfType(fNodes, EN::FILTER, true);
   if (fNodes.empty()) {
     // no filters present
@@ -6056,8 +6053,8 @@ void arangodb::aql::removeFiltersCoveredByTraversal(Optimizer* opt,
 void arangodb::aql::removeTraversalPathVariable(Optimizer* opt,
                                                 std::unique_ptr<ExecutionPlan> plan,
                                                 OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> tNodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> tNodesStorage;
+  auto& tNodes = tNodesStorage.vector();
   plan->findNodesOfType(tNodes, EN::TRAVERSAL, true);
 
   bool modified = false;
@@ -6092,8 +6089,8 @@ void arangodb::aql::removeTraversalPathVariable(Optimizer* opt,
 ///   RETURN x.a
 void arangodb::aql::inlineSubqueriesRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
                                          OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::SUBQUERY, true);
 
   if (nodes.empty()) {
@@ -6919,8 +6916,8 @@ static bool applyGeoOptimization(ExecutionPlan* plan, LimitNode* ln,
 
 void arangodb::aql::geoIndexRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
                                  OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   bool mod = false;
 
   plan->findNodesOfType(nodes, EN::ENUMERATE_COLLECTION, true);
@@ -7041,8 +7038,8 @@ void arangodb::aql::sortLimitRule(Optimizer* opt, std::unique_ptr<ExecutionPlan>
     return;
   }
 
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> limitNodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> limitNodesStorage;
+  auto& limitNodes = limitNodesStorage.vector();
 
   plan->findNodesOfType(limitNodes, EN::LIMIT, true);
   for (ExecutionNode* node : limitNodes) {
@@ -7083,8 +7080,8 @@ void arangodb::aql::optimizeSubqueriesRule(Optimizer* opt,
                                            OptimizerRule const& rule) {
   bool modified = false;
 
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::CALCULATION, true);
 
   // value type is {limit value, referenced by, used for counting}
@@ -7276,8 +7273,8 @@ void arangodb::aql::moveFiltersIntoEnumerateRule(Optimizer* opt,
                                                  OptimizerRule const& rule) {
   bool modified = false;
 
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, ::moveFilterIntoEnumerateTypes, true);
 
   VarSet found;
@@ -7457,8 +7454,8 @@ void arangodb::aql::optimizeCountRule(Optimizer* opt, std::unique_ptr<ExecutionP
     return;
   }
 
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::CALCULATION, true);
 
   VarSet vars;
@@ -7760,8 +7757,8 @@ void arangodb::aql::parallelizeGatherRule(Optimizer* opt,
   // thread-safe here. the only case in which we currently can employ parallelization
   // is when there is only a single GatherNode. all other restrictions for
   // parallelization (e.g. no DistributeNodes around) still apply.
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::GATHER, true);
 
   if (nodes.size() == 1 && !plan->contains(EN::DISTRIBUTE) && !plan->contains(EN::SCATTER)) {
@@ -7833,8 +7830,8 @@ void arangodb::aql::enableAsyncPrefetching(ExecutionPlan& plan) {
 }
 
 void arangodb::aql::enableReadOwnWritesForUpsertSubquery(ExecutionPlan& plan) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan.findNodesOfType(nodes, EN::UPSERT, true);
 
   // An upsert is roughly translated to the following:
@@ -8042,8 +8039,8 @@ void arangodb::aql::spliceSubqueriesRule(Optimizer* opt, std::unique_ptr<Executi
                                          OptimizerRule const& rule) {
   bool modified = false;
 
-  containers::SmallVector<SubqueryNode*>::allocator_type::arena_type a;
-  containers::SmallVector<SubqueryNode*> subqueryNodes{a};
+  ::arangodb::containers::SmallVectorWithArena<SubqueryNode*> subqueryNodesStorage;
+  auto& subqueryNodes = subqueryNodesStorage.vector();
   findSubqueriesSuitableForSplicing(*plan, subqueryNodes);
 
   // Note that we rely on the order `subqueryNodes` in the sense that, for
@@ -8172,8 +8169,8 @@ void arangodb::aql::spliceSubqueriesRule(Optimizer* opt, std::unique_ptr<Executi
 void arangodb::aql::decayUnnecessarySortedGather(Optimizer* opt,
                                                  std::unique_ptr<ExecutionPlan> plan,
                                                  OptimizerRule const& rule) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan->findNodesOfType(nodes, EN::GATHER, true);
 
   bool modified = false;
@@ -8199,8 +8196,8 @@ void arangodb::aql::decayUnnecessarySortedGather(Optimizer* opt,
 }
 
 void arangodb::aql::insertDistributeInputCalculation(ExecutionPlan& plan) {
-  ::arangodb::containers::SmallVector<ExecutionNode*>::allocator_type::arena_type a;
-  ::arangodb::containers::SmallVector<ExecutionNode*> nodes{a};
+  ::arangodb::containers::SmallVectorWithArena<ExecutionNode*> nodesStorage;
+  auto& nodes = nodesStorage.vector();
   plan.findNodesOfType(nodes, ExecutionNode::DISTRIBUTE, true);
 
   for (auto const& n : nodes) {
