@@ -31,12 +31,12 @@
 
 namespace arangodb::cache {
 
-BucketState::BucketState() : _state(0) {}
+BucketState::BucketState() noexcept : _state(0) {}
 
-BucketState::BucketState(BucketState const& other)
+BucketState::BucketState(BucketState const& other) noexcept 
     : _state(other._state.load()) {}
 
-BucketState& BucketState::operator=(BucketState const& other) {
+BucketState& BucketState::operator=(BucketState const& other) noexcept {
   if (this != &other) {
     _state = other._state.load();
   }
@@ -44,11 +44,11 @@ BucketState& BucketState::operator=(BucketState const& other) {
   return *this;
 }
 
-bool BucketState::isLocked() const {
+bool BucketState::isLocked() const noexcept {
   return ((_state.load() & static_cast<std::uint32_t>(Flag::locked)) > 0);
 }
 
-bool BucketState::lock(std::uint64_t maxTries, BucketState::CallbackType cb) {
+bool BucketState::lock(std::uint64_t maxTries) noexcept {
   uint64_t attempt = 0;
   while (attempt < maxTries) {
     // expect unlocked, but need to preserve migrating status
@@ -60,7 +60,6 @@ bool BucketState::lock(std::uint64_t maxTries, BucketState::CallbackType cb) {
       bool success = _state.compare_exchange_strong(expected, desired, std::memory_order_acq_rel,
                                                     std::memory_order_relaxed);
       if (success) {
-        cb();
         return true;
       }
     }
@@ -72,28 +71,28 @@ bool BucketState::lock(std::uint64_t maxTries, BucketState::CallbackType cb) {
   return false;
 }
 
-void BucketState::unlock() {
+void BucketState::unlock() noexcept {
   TRI_ASSERT(isLocked());
   _state.fetch_and(~static_cast<std::uint32_t>(Flag::locked), std::memory_order_release);
 }
 
-bool BucketState::isSet(BucketState::Flag flag) const {
+bool BucketState::isSet(BucketState::Flag flag) const noexcept {
   TRI_ASSERT(isLocked());
   return ((_state.load() & static_cast<std::uint32_t>(flag)) > 0);
 }
 
-bool BucketState::isSet(BucketState::Flag flag1, BucketState::Flag flag2) const {
+bool BucketState::isSet(BucketState::Flag flag1, BucketState::Flag flag2) const noexcept {
   TRI_ASSERT(isLocked());
   return ((_state.load() & (static_cast<std::uint32_t>(flag1) |
                             static_cast<std::uint32_t>(flag2))) > 0);
 }
 
-void BucketState::toggleFlag(BucketState::Flag flag) {
+void BucketState::toggleFlag(BucketState::Flag flag) noexcept {
   TRI_ASSERT(isLocked());
   _state ^= static_cast<std::uint32_t>(flag);
 }
 
-void BucketState::clear() {
+void BucketState::clear() noexcept {
   TRI_ASSERT(isLocked());
   _state = static_cast<std::uint32_t>(Flag::locked);
 }
