@@ -27,6 +27,7 @@
 #include <cstdint>
 #include <functional>
 #include <limits>
+#include <type_traits>
 
 namespace arangodb {
 namespace cache {
@@ -43,8 +44,6 @@ namespace cache {
 /// via an enum and must correspond to exactly one set bit.
 ////////////////////////////////////////////////////////////////////////////////
 struct BucketState {
-  typedef std::function<void()> CallbackType;
-
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Flags which can be queried or toggled to reflect state.
   ///
@@ -70,22 +69,22 @@ struct BucketState {
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Initializes state with no flags set and unlocked
   //////////////////////////////////////////////////////////////////////////////
-  BucketState();
+  BucketState() noexcept;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Initializes state to match another
   //////////////////////////////////////////////////////////////////////////////
-  BucketState(BucketState const& other);
+  BucketState(BucketState const& other) noexcept;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Initializes state to match another
   //////////////////////////////////////////////////////////////////////////////
-  BucketState& operator=(BucketState const& other);
+  BucketState& operator=(BucketState const& other) noexcept;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Checks if state is locked.
   //////////////////////////////////////////////////////////////////////////////
-  bool isLocked() const;
+  bool isLocked() const noexcept;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Used to lock state. Returns true if locked, false otherwise.
@@ -96,29 +95,38 @@ struct BucketState {
   /// locked or not. The optional second parameter is a function which will be
   /// called upon successfully locking the state.
   //////////////////////////////////////////////////////////////////////////////
-  bool lock(std::uint64_t maxTries = std::numeric_limits<std::uint64_t>::max(),
-            BucketState::CallbackType cb = []() -> void {});
+  bool lock(std::uint64_t maxTries = std::numeric_limits<std::uint64_t>::max()) noexcept;
+
+  template<typename F>
+  bool lock(std::uint64_t maxTries, F&& cb) noexcept {
+    static_assert(std::is_nothrow_invocable_r_v<void, F>);
+    bool success = this->lock(maxTries);
+    if (success) {
+      std::invoke(std::forward<F>(cb));
+    }
+    return success;
+  }
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Unlocks the state. Requires state to be locked.
   //////////////////////////////////////////////////////////////////////////////
-  void unlock();
+  void unlock() noexcept;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Checks whether the given flag is set. Requires state to be locked.
   //////////////////////////////////////////////////////////////////////////////
-  bool isSet(BucketState::Flag flag) const;
-  bool isSet(BucketState::Flag flag1, BucketState::Flag flag2) const;
+  bool isSet(BucketState::Flag flag) const noexcept;
+  bool isSet(BucketState::Flag flag1, BucketState::Flag flag2) const noexcept;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Toggles the given flag. Requires state to be locked.
   //////////////////////////////////////////////////////////////////////////////
-  void toggleFlag(BucketState::Flag flag);
+  void toggleFlag(BucketState::Flag flag) noexcept;
 
   //////////////////////////////////////////////////////////////////////////////
   /// @brief Unsets all flags besides Flag::locked. Requires state to be locked.
   //////////////////////////////////////////////////////////////////////////////
-  void clear();
+  void clear() noexcept;
 
  private:
   std::atomic<std::uint32_t> _state;
