@@ -110,7 +110,7 @@ void CacheManagerFeature::start() {
   };
 
   SharedPRNGFeature& sharedPRNG = server().getFeature<SharedPRNGFeature>();
-  _manager = std::make_unique<Manager>(sharedPRNG, postFn, _cacheSize);
+  _manager = std::make_unique<Manager>(sharedPRNG, std::move(postFn), _cacheSize);
 
   _rebalancer = std::make_unique<CacheRebalancerThread>(server(), _manager.get(), _rebalancingInterval);
   _rebalancer->start();
