@@ -88,7 +88,7 @@ Manager::Manager(SharedPRNGFeature& sharedPRNG,
       _spareTableAllocation(0),
       _globalAllocation(_fixedAllocation),
       _transactions(),
-      _schedulerPost(schedulerPost),
+      _schedulerPost(std::move(schedulerPost)),
       _resizeAttempt(0),
       _outstandingTasks(0),
       _rebalancingTasks(0),
@@ -591,9 +591,11 @@ void Manager::shrinkOvergrownCaches(Manager::TaskEnvironment environment) {
 
 void Manager::freeUnusedTables() {
   TRI_ASSERT(_lock.isLockedWrite());
-  for (std::size_t i = 0; i < 32; i++) {
+  constexpr std::size_t tableEntries = std::tuple_size<decltype(_tables)>::value;
+
+  for (std::size_t i = 0; i < tableEntries; i++) {
     while (!_tables[i].empty()) {
-      auto table = _tables[i].top();
+      auto table = std::move(_tables[i].top());
       _globalAllocation -= table->memoryUsage();
       TRI_ASSERT(_globalAllocation >= _fixedAllocation);
       _tables[i].pop();
@@ -673,6 +675,7 @@ std::shared_ptr<Table> Manager::leaseTable(std::uint32_t logSize) {
   TRI_ASSERT(_lock.isLockedWrite());
 
   std::shared_ptr<Table> table;
+  TRI_ASSERT(_tables.size() >= logSize);
   if (_tables[logSize].empty()) {
     if (increaseAllowed(Table::allocationSize(logSize), true)) {
       try {
@@ -685,6 +688,7 @@ std::shared_ptr<Table> Manager::leaseTable(std::uint32_t logSize) {
     }
   } else {
     table = _tables[logSize].top();
+    TRI_ASSERT(table != nullptr);
     _spareTableAllocation -= table->memoryUsage();
     _tables[logSize].pop();
   }
@@ -693,7 +697,7 @@ std::shared_ptr<Table> Manager::leaseTable(std::uint32_t logSize) {
 }
 
 void Manager::reclaimTable(std::shared_ptr<Table> table, bool internal) {
-  TRI_ASSERT(table.get() != nullptr);
+  TRI_ASSERT(table != nullptr);
   SpinLocker guard(SpinLocker::Mode::Write, _lock, !internal);
 
   std::uint32_t logSize = table->logSize();
