@@ -49,9 +49,7 @@ Finding TransactionalCache::find(void const* key, std::uint32_t keySize) {
   Finding result;
   std::uint32_t hash = hashKey(key, keySize);
 
-  Result status;
-  Table::BucketLocker guard;
-  std::tie(status, guard) = getBucket(hash, Cache::triesFast, false);
+  auto [status, guard] = getBucket(hash, Cache::triesFast, false);
   if (status.fail()) {
     result.reportError(status);
     return result;
@@ -121,6 +119,7 @@ Result TransactionalCache::insert(CachedValue* value) {
           }
           bucket.insert(hash, value);
           if (!eviction) {
+            TRI_ASSERT(source != nullptr);
             maybeMigrate = source->slotFilled();
           }
           maybeMigrate |= reportInsert(eviction);
@@ -135,6 +134,7 @@ Result TransactionalCache::insert(CachedValue* value) {
   }
 
   if (maybeMigrate) {
+    TRI_ASSERT(source != nullptr);
     requestMigrate(source->idealSize());  // let function do the hard work
   }
 
@@ -168,11 +168,13 @@ Result TransactionalCache::remove(void const* key, std::uint32_t keySize) {
       }
 
       freeValue(candidate);
+      TRI_ASSERT(source != nullptr);
       maybeMigrate = source->slotEmptied();
     }
   }
 
   if (maybeMigrate) {
+    TRI_ASSERT(source != nullptr);
     requestMigrate(source->idealSize());
   }
 
@@ -207,11 +209,13 @@ Result TransactionalCache::banish(void const* key, std::uint32_t keySize) {
       }
 
       freeValue(candidate);
+      TRI_ASSERT(source != nullptr);
       maybeMigrate = source->slotEmptied();
     }
   }
 
   if (maybeMigrate) {
+    TRI_ASSERT(source != nullptr);
     requestMigrate(source->idealSize());
   }
 
@@ -274,7 +278,7 @@ uint64_t TransactionalCache::freeMemoryFrom(std::uint32_t hash) {
     }
   }
 
-  cache::Table* table = _table.load(std::memory_order_relaxed);
+  std::shared_ptr<cache::Table> table = this->table();
   if (table) {
     std::int32_t size = table->idealSize();
     if (maybeMigrate) {
@@ -291,7 +295,9 @@ void TransactionalCache::migrateBucket(void* sourcePtr,
   std::uint64_t term = _manager->_transactions.term();
 
   // lock current bucket
-  Table::BucketLocker sourceGuard(sourcePtr, _table.load(), Cache::triesGuarantee);
+  std::shared_ptr<Table> table = this->table();
+
+  Table::BucketLocker sourceGuard(sourcePtr, table.get(), Cache::triesGuarantee);
   TransactionalBucket& source = sourceGuard.bucket<TransactionalBucket>();
   term = std::max(term, source._banishTerm);
 
@@ -320,6 +326,8 @@ void TransactionalCache::migrateBucket(void* sourcePtr,
         return true;
       });
     } else {
+      std::uint64_t totalSize = 0;
+      std::uint64_t emptied = 0;
       for (std::size_t j = 0; j < TransactionalBucket::slotsBanish; j++) {
         std::uint32_t hash = source._banishHashes[j];
         if (hash != 0) {
@@ -329,12 +337,14 @@ void TransactionalCache::migrateBucket(void* sourcePtr,
           if (candidate != nullptr) {
             std::uint64_t size = candidate->size();
             freeValue(candidate);
-            reclaimMemory(size);
-            newTable->slotEmptied();
+            totalSize += size;
+            ++emptied;
           }
           source._banishHashes[j] = 0;
         }
       }
+      reclaimMemory(totalSize);
+      newTable->slotsEmptied(emptied);
     }
 
     // migrate actual values
@@ -389,7 +399,7 @@ std::tuple<Result, Table::BucketLocker> TransactionalCache::getBucket(
   Result status;
   Table::BucketLocker guard;
 
-  Table* table = _table.load(std::memory_order_relaxed);
+  std::shared_ptr<Table> table = this->table();
   if (isShutdown() || table == nullptr) {
     status.reset(TRI_ERROR_SHUTTING_DOWN);
     return std::make_tuple(std::move(status), std::move(guard));
