@@ -63,7 +63,7 @@ typename std::enable_if<std::is_base_of<std::exception, E>::value, Future<T>>::t
 }
 
 // makeFutureWith(Future<T>()) -> Future<T>
-template <typename F, typename R = std::result_of_t<F()>>
+template <typename F, typename R = std::invoke_result_t<F>>
 typename std::enable_if<isFuture<R>::value, R>::type makeFutureWith(F&& func) {
   using InnerType = typename isFuture<R>::inner;
   try {
@@ -75,7 +75,7 @@ typename std::enable_if<isFuture<R>::value, R>::type makeFutureWith(F&& func) {
 
 // makeFutureWith(T()) -> Future<T>
 // makeFutureWith(void()) -> Future<Unit>
-template <typename F, typename R = std::result_of_t<F()>>
+template <typename F, typename R = std::invoke_result_t<F>>
 typename std::enable_if<!isFuture<R>::value, Future<R>>::type makeFutureWith(F&& func) {
   return makeFuture<R>(
       makeTryWith([&func]() mutable { return std::forward<F>(func)(); }));
