@@ -36,9 +36,9 @@ TEST(CacheBucketStateTest, test_lock_methods) {
 
   std::uint32_t outsideBucketState = 0;
 
-  auto cb1 = [&outsideBucketState]() -> void { outsideBucketState = 1; };
+  auto cb1 = [&outsideBucketState]() noexcept -> void { outsideBucketState = 1; };
 
-  auto cb2 = [&outsideBucketState]() -> void { outsideBucketState = 2; };
+  auto cb2 = [&outsideBucketState]() noexcept -> void { outsideBucketState = 2; };
 
   // check lock without contention
   ASSERT_FALSE(state.isLocked());
